requirements.txt
discord.py
Flask
aiohttp
python-dotenv
gunicorn
Flask-CORS
re